Split data.rs into smaller files
