+++ talpid-routing/src/unix/macos/data/message_type.rs
@@ -1,0 +1,39 @@
+bitflags::bitflags! {
+    /// Types of routing messages
+    /// See https://www.manpagez.com/man/4/route/.
+    #[derive(Debug)]
+    pub struct MessageType: i32 {
+        /// Add Route
+        const RTM_ADD         = libc::RTM_ADD;
+        /// Delete Route
+        const RTM_DELETE      = libc::RTM_DELETE;
+        /// Change Metrics or flags
+        const RTM_CHANGE      = libc::RTM_CHANGE;
+        /// Report Metrics
+        const RTM_GET         = libc::RTM_GET;
+        /// RTM_LOSING is no longer generated by and is deprecated
+        const RTM_LOSING      = libc::RTM_LOSING;
+        /// Told to use different route
+        const RTM_REDIRECT    = libc::RTM_REDIRECT;
+        /// Lookup failed on this address
+        const RTM_MISS        = libc::RTM_MISS;
+        /// fix specified metrics
+        const RTM_LOCK        = libc::RTM_LOCK;
+        /// caused by SIOCADDRT
+        const RTM_OLDADD      = libc::RTM_OLDADD;
+        /// caused by SIOCDELRT
+        const RTM_OLDDEL      = libc::RTM_OLDDEL;
+        /// req to resolve dst to LL addr
+        const RTM_RESOLVE     = libc::RTM_RESOLVE;
+        /// address being added to iface
+        const RTM_NEWADDR     = libc::RTM_NEWADDR;
+        /// address being removed from iface
+        const RTM_DELADDR     = libc::RTM_DELADDR;
+        /// iface going up/down etc.
+        const RTM_IFINFO      = libc::RTM_IFINFO;
+        /// mcast group membership being added to if
+        const RTM_NEWMADDR    = libc::RTM_NEWMADDR;
+        /// mcast group membership being deleted
+        const RTM_DELMADDR    = libc::RTM_DELMADDR;
+    }
+}
